complete Lesson15 kadai-1

+++ app/Http/Controllers/TasksController.php
@@ -15,11 +15,40 @@ class TasksController extends Controller
      */
     public function index()
     {
+        /*
         $tasks = Task::all();
 
         return view('tasks.index', [
             'tasks' => $tasks,
         ]);
+        */
+        
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasklists = $user->tasklists()->orderBy('created_at', 'desc')->paginate(10);
+            
+            $data = [
+                'user' => $user,
+                'tasks' => $tasklists,
+            ];
+            
+            //dd($data);
+            
+            //return view('welcome', $data);
+            //return view('tasks.index', $data);
+            
+/*        } else {
+            $data = [];
+            
+            return view('welcome', $data);
+*/
+        }
+        
+        //dd($data);
+        
+        return view('welcome', $data);
+        
     }
 
     /**
@@ -48,13 +77,36 @@ class TasksController extends Controller
             'status' => 'required|max:10',
             'content' => 'required|max:191',
         ]);
+        
+        $request->user()->tasklists()->create([
+            'status'  => $request->status,
+            'content' => $request->content,
+        ]);
 
+        return redirect('/');
+        
+        /*
+        $user = \Auth::user();
+        $tasklists = $user->tasklists()->orderBy('created_at', 'desc')->paginate(10);
+            
+        $data = [
+            'user' => $user,
+            'tasks' => $tasklists,
+        ];
+
+        //dd($data);
+
+        return view('welcome',$data);
+        */
+        
+        /*
         $task = new Task;
         $task->status = $request->status;
         $task->content = $request->content;
         $task->save();
 
         return redirect('/');
+        */
     }
 
     /**
@@ -105,8 +157,36 @@ class TasksController extends Controller
         $task->status = $request->status;
         $task->content = $request->content;
         $task->save();
+        
+        return redirect('/');
+        
+        /*
+        $request->user()->tasklists()->save([
+            'status'  => $request->status,
+            'content' => $request->content,
+        ]);
+
+        $user = \Auth::user();
+        $tasklists = $user->tasklists()->orderBy('created_at', 'desc')->paginate(10);
+        
+        $data = [
+            'user' => $user,
+            'tasks' => $tasklists,
+        ];
+
+        //dd($data);
+
+        return view('welcome',$data);
+        */
+        
+        /*
+        $task = Task::find($id);
+        $task->status = $request->status;
+        $task->content = $request->content;
+        $task->save();
 
         return redirect('/');
+        */
     }
 
     /**
@@ -118,8 +198,29 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::find($id);
-        $task->delete();
-
+        
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
+        
         return redirect('/');
+        
+        /*
+        $user = \Auth::user();
+        $tasklists = $user->tasklists()->orderBy('created_at', 'desc')->paginate(10);
+        
+        $data = [
+            'user' => $user,
+            'tasks' => $tasklists,
+        ];
+
+        //dd($data);
+
+        return view('welcome',$data);
+        */
+        
+        //$task->delete();
+
+        //return redirect('/');
     }
 }
